Added config to API
