+++ Gems/Pointcloud/Code/Source/Tools/Components/PointcloudComponentController.h
@@ -14,26 +14,10 @@
 #include <AzCore/Math/Aabb.h>
 #include <Pointcloud/PointcloudConfigurationBus.h>
 #include <Pointcloud/PointcloudFeatureProcessorInterface.h>
+#include <Pointcloud/PointcloudComponentConfig.h>
 
 namespace Pointcloud
 {
-    class PointcloudComponentConfig final : public AZ::ComponentConfig
-    {
-    public:
-        AZ_RTTI(PointcloudComponentConfig, "{3ae848a0-3cd0-439e-bafe-db0270caae47}");
-
-        PointcloudComponentConfig() = default;
-        ~PointcloudComponentConfig() = default;
-
-        static void Reflect(AZ::ReflectContext* context);
-
-        AZ::EntityId m_editorEntityId;
-        float m_pointSize = 1.0f;
-        PointcloudFeatureProcessorInterface::PointcloudHandle m_pointcloudHandle =
-            PointcloudFeatureProcessorInterface::InvalidPointcloudHandle;
-        AZ::Data::Asset<PointcloudAsset> m_pointcloudAsset = {};
-    };
-
     class PointcloudComponentController
         : public PointcloudConfigurationBus::Handler
         , private AZ::TransformNotificationBus::Handler
